// NBA/src/app/interfaces/teamPlayers.interface.ts
// Generated by https://quicktype.io

export interface TeamInfoResponse {
    _internal: Internal;
    league:    League;
}

export interface Internal {
    pubDateTime:             string;
    igorPath:                string;
    xslt:                    string;
    xsltForceRecompile:      string;
    xsltInCache:             string;
    xsltCompileTimeMillis:   string;
    xsltTransformTimeMillis: string;
    consolidatedDomKey:      string;
    endToEndTimeMillis:      string;
}

export interface League {
    standard:   Teaminfo;
    africa:     Teaminfo;
    sacramento: Teaminfo;
    vegas:      Teaminfo;
    utah:       Teaminfo;
}

export interface Teaminfo {
    teamId:  string;
    players: PlayerElement[];
}

export interface PlayerElement {
    personId: string;
}